+++ Magic/src/main/java/com/elmakers/mine/bukkit/spell/builtin/GotoSpell.java
@@ -80,7 +80,7 @@ public class GotoSpell extends UndoableSpell
             getCurrentTarget().setEntity(targetEntity);
             registerModified(player);
             teleportTo(player, targetEntity);
-            castMessage(getMessage("cast_to_player").replace("$target", controller.getEntityDisplayName(targetEntity)));
+            castMessageKey("cast_to_player", getMessage("cast_to_player").replace("$target", controller.getEntityDisplayName(targetEntity)));
 
             return SpellResult.CAST;
         }
@@ -116,7 +116,7 @@ public class GotoSpell extends UndoableSpell
             getCurrentTarget().setEntity(targetEntity);
             registerModified(player);
             teleportTo(player, targetEntity);
-            castMessage(getMessage("cast_to_player").replace("$target", controller.getEntityDisplayName(targetEntity)));
+            castMessageKey("cast_to_player", getMessage("cast_to_player").replace("$target", controller.getEntityDisplayName(targetEntity)));
             releaseTarget();
             registerForUndo();
             return SpellResult.CAST;